refactor: web3auth config

Took 4 minutes
