+++ src/app/explore/page.tsx
@@ -1,18 +1,19 @@
 'use client'
 import Image from 'next/image'
-import React, { useEffect, useState } from 'react'
-import { PiGoogleLogoBold } from 'react-icons/pi'
-import { Web3AuthNoModal } from '@web3auth/no-modal'
-import type { IProvider } from '@web3auth/base'
-import { WALLET_ADAPTERS } from '@web3auth/base'
-import { EthereumPrivateKeyProvider } from '@web3auth/ethereum-provider'
-import { OpenloginAdapter } from '@web3auth/openlogin-adapter'
-import { useRouter } from 'next/navigation'
-import { useUserContext } from '../userContext'
-import { create_user } from '@/lib/prismaUtils'
-import { generateFromEmail } from 'unique-username-generator'
-import { AvatarGenerator } from 'random-avatar-generator'
-import { AppName, Web3AuthConfig, Web3AuthEthPrivateKeyProviderConfig } from '@enums/app'
+import React, {useEffect, useState} from 'react'
+import {PiGoogleLogoBold} from 'react-icons/pi'
+import {Web3AuthNoModal} from '@web3auth/no-modal'
+import type {IProvider} from '@web3auth/base'
+import {WALLET_ADAPTERS} from '@web3auth/base'
+import {EthereumPrivateKeyProvider} from '@web3auth/ethereum-provider'
+import {OpenloginAdapter} from '@web3auth/openlogin-adapter'
+import {useRouter} from 'next/navigation'
+import {useUserContext} from '../userContext'
+import {create_user} from '@/lib/prismaUtils'
+import {generateFromEmail} from 'unique-username-generator'
+import {AvatarGenerator} from 'random-avatar-generator'
+import {AppName} from '../../../config/app'
+import {Web3AuthConfig, Web3AuthEthPrivateKeyProviderConfig} from "../../../config/authConfig";
 
 const Page = () => {
   const [view, setView] = useState<boolean>(false)
@@ -161,7 +162,7 @@ const Page = () => {
         <div className="w-[80%] lg:w-[30%] h-[80%] ">
           <div className="h-[10%] ">
             <h1 className="font-logirentBold text-primary_1 text-center font-bold text-4xl">
-              {' '} 
+              {' '}
               Explore
             </h1>
           </div>
