+++ src/app/providers.tsx
@@ -1,16 +1,14 @@
 'use client'
 
 import * as React from 'react'
-import {
-  getDefaultWallets,
-  midnightTheme,
-  RainbowKitProvider,
-} from '@rainbow-me/rainbowkit'
+import {getDefaultWallets, midnightTheme, RainbowKitProvider,} from '@rainbow-me/rainbowkit'
 
-import { configureChains, createConfig, WagmiConfig } from 'wagmi'
-import { filecoinCalibration, mainnet, polygon } from 'wagmi/chains'
-import { publicProvider } from 'wagmi/providers/public'
-import { AppName, WagmiProjectId } from '@enums/app'
+import {configureChains, createConfig, WagmiConfig} from 'wagmi'
+import {filecoinCalibration, mainnet, polygon} from 'wagmi/chains'
+import {publicProvider} from 'wagmi/providers/public'
+import {AppName} from '../../config/app'
+
+import {WagmiProjectId} from "../../config/authConfig";
 // const lilypad = {
 //   id: 1337,
 //   name: "Lilypad Lalechuza testnet",
